Adding the ability to POST JSON data
Added a -j <JSON> option flag. If set will POST JSON data to URL.
POSTing JSON data to elasticsearch allows us to query facets of the data in the indexes
e.g. get a count of the number of 404's in the past 5 mins
TODO Add a check to ensure JSON data is being passed if the option is set.